+++ metrics-servlets/src/main/java/com/codahale/metrics/servlets/MetricsServlet.java
@@ -5,8 +5,7 @@ import com.fasterxml.jackson.databind.ObjectWriter;
 import com.codahale.metrics.MetricRegistry;
 import com.codahale.metrics.json.MetricsModule;
 
-import javax.servlet.ServletConfig;
-import javax.servlet.ServletException;
+import javax.servlet.*;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -15,53 +14,104 @@ import java.io.OutputStream;
 import java.util.Locale;
 import java.util.concurrent.TimeUnit;
 
+/**
+ * A servlet which returns the metrics in a given registry as an {@code application/json} response.
+ */
 public class MetricsServlet extends HttpServlet {
+    /**
+     * An abstract {@link ServletContextListener} which allows you to programmatically inject the
+     * {@link MetricRegistry}, rate and duration units, and allowed origin for
+     * {@link MetricsServlet}.
+     */
+    public static abstract class ContextListener implements ServletContextListener {
+        /**
+         * Returns the {@link MetricRegistry} to inject into the servlet context.
+         */
+        protected abstract MetricRegistry getMetricRegistry();
+
+        /**
+         * Returns the {@link TimeUnit} to which rates should be converted, or {@code null} if the
+         * default should be used.
+         */
+        protected TimeUnit getRateUnit() {
+            // use the default
+            return null;
+        }
+
+        /**
+         * Returns the {@link TimeUnit} to which durations should be converted, or {@code null} if
+         * the default should be used.
+         */
+        protected TimeUnit getDurationUnit() {
+            // use the default
+            return null;
+        }
+
+        /**
+         * Returns the {@code Access-Control-Allow-Origin} header value, if any.
+         */
+        protected String getAllowedOrigin() {
+            // use the default
+            return null;
+        }
+
+        @Override
+        public void contextInitialized(ServletContextEvent event) {
+            final ServletContext context = event.getServletContext();
+            context.setAttribute(METRICS_REGISTRY, getMetricRegistry());
+            context.setAttribute(RATE_UNIT, getRateUnit());
+            context.setAttribute(DURATION_UNIT, getDurationUnit());
+            context.setAttribute(ALLOWED_ORIGIN, getAllowedOrigin());
+        }
+
+        @Override
+        public void contextDestroyed(ServletContextEvent event) {
+            // no-op
+        }
+    }
+
     public static final String RATE_UNIT = MetricsServlet.class.getCanonicalName() + ".rateUnit";
     public static final String DURATION_UNIT = MetricsServlet.class.getCanonicalName() + ".durationUnit";
     public static final String SHOW_SAMPLES = MetricsServlet.class.getCanonicalName() + ".showSamples";
     public static final String METRICS_REGISTRY = MetricsServlet.class.getCanonicalName() + ".registry";
-    public static final String ALLOWED_ORIGINS = MetricsServlet.class.getCanonicalName() + ".corsAllowedOrigins";
+    public static final String ALLOWED_ORIGIN = MetricsServlet.class.getCanonicalName() + ".allowedOrigin";
 
     private static final long serialVersionUID = 1049773947734939602L;
     private static final String CONTENT_TYPE = "application/json";
 
-    private String allowedOrigins;
+    private String allowedOrigin;
     private transient MetricRegistry registry;
     private transient ObjectMapper mapper;
 
     @Override
     public void init(ServletConfig config) throws ServletException {
-        final Object registryAttr = config.getServletContext().getAttribute(METRICS_REGISTRY);
+        final ServletContext context = config.getServletContext();
+
+        final Object registryAttr = context.getAttribute(METRICS_REGISTRY);
         if (registryAttr instanceof MetricRegistry) {
             this.registry = (MetricRegistry) registryAttr;
         } else {
             throw new ServletException("Couldn't find a MetricRegistry instance.");
         }
 
-        final TimeUnit rateUnit = parseTimeUnit(config.getServletContext()
-                                                      .getInitParameter(RATE_UNIT),
+        final TimeUnit rateUnit = parseTimeUnit(context.getInitParameter(RATE_UNIT),
                                                 TimeUnit.SECONDS);
-
-        final TimeUnit durationUnit = parseTimeUnit(config.getServletContext()
-                                                          .getInitParameter(DURATION_UNIT),
+        final TimeUnit durationUnit = parseTimeUnit(context.getInitParameter(DURATION_UNIT),
                                                     TimeUnit.SECONDS);
-
-        final boolean showSamples = Boolean.parseBoolean(config.getServletContext()
-                                                               .getInitParameter(SHOW_SAMPLES));
-
-        this.allowedOrigins = config.getInitParameter(ALLOWED_ORIGINS);
-
+        final boolean showSamples = Boolean.parseBoolean(context.getInitParameter(SHOW_SAMPLES));
         this.mapper = new ObjectMapper().registerModule(new MetricsModule(rateUnit,
                                                                           durationUnit,
                                                                           showSamples));
+
+        this.allowedOrigin = config.getInitParameter(ALLOWED_ORIGIN);
     }
 
     @Override
     protected void doGet(HttpServletRequest req,
                          HttpServletResponse resp) throws ServletException, IOException {
         resp.setContentType(CONTENT_TYPE);
-        if (allowedOrigins != null) {
-            resp.setHeader("Access-Control-Allow-Origin", allowedOrigins);
+        if (allowedOrigin != null) {
+            resp.setHeader("Access-Control-Allow-Origin", allowedOrigin);
         }
         resp.setHeader("Cache-Control", "must-revalidate,no-cache,no-store");
         resp.setStatus(HttpServletResponse.SC_OK);